refactor: ajusta lógica para filtrar dados antes de fazer a requisição
